Improve the parser and add a 'Start' frame.
The parser can handle other trace messages existing and is less specific
about the 'garbage' before and after the source and trace events.

A 'Start' frame is at the bottom of the call stack so if we return to it
we know the entry point was returned from successfully and we can exit.